+++ server/configuration.go
@@ -18,6 +18,8 @@ import (
 // If you add non-reference types to your configuration struct, be sure to rewrite Clone as a deep
 // copy appropriate for your types.
 type configuration struct {
+	TeamFilesEnabled      bool
+	AdminTeamFilesEnabled bool
 }
 
 // Clone shallow copies the configuration. Your implementation may require a deep copy if