+++ webapp/src/reducer.js
@@ -1,6 +1,6 @@
 import {combineReducers} from 'redux';
 
-import {OPEN_ROOT_MODAL, CLOSE_ROOT_MODAL, LOAD_FILES, SET_ERROR} from './action_types';
+import {OPEN_ROOT_MODAL, CLOSE_ROOT_MODAL, LOAD_FILES, SET_ERROR, SET_CONFIG} from './action_types';
 
 const rootModalVisible = (state = false, action) => {
     switch (action.type) {
@@ -16,7 +16,18 @@ const rootModalVisible = (state = false, action) => {
 const isModalForTeam = (state = false, action) => {
     switch(action.type) {
         case OPEN_ROOT_MODAL:
-            return action.payload;
+            return action.isModalForTeam || false;
+        case CLOSE_ROOT_MODAL:
+            return false;
+        default:
+            return state;
+    }
+}
+
+const isTeamModalForAdmin = (state = false, action) => {
+    switch(action.type) {
+        case OPEN_ROOT_MODAL:
+            return action.isTeamModalForAdmin || false;
         case CLOSE_ROOT_MODAL:
             return false;
         default:
@@ -42,9 +53,20 @@ const error = (state = null, action) => {
     }
 }
 
+const config = (state = null, action) => {
+    switch(action.type) {
+        case SET_CONFIG:
+            return action.payload;
+        default:
+            return state;
+    }
+}
+
 export default combineReducers({
     rootModalVisible,
     isModalForTeam,
+    isTeamModalForAdmin,
     files,
-    error
+    error,
+    config
 });