+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.7.0 - 2020-01-19
+### Added
+- Lists of all files in the team. Optionally for admins also a list of all files even from channels where the admin is not a member. Both of those list are opt-in: have to be enabled in pluggin settings.
+- Disabled "Get public link" button when the public links are disabled in the admin console
+
 ## 0.6.0 - 2020-01-12
 
 ### Breaking changes